feat(skolengo): add Skolengo authentification flow
At present, we're using Webviews for compatibility with Expo Go. Another means may be implemented later with “intents” (skoapp-prod://) but will have to detect whether we're using a prebuild or Expo-go. In addition, the current script logs in and attempts to obtain user data. The account is not yet stored, and the lack of access before the start of the school year means we can't make any further progress on the data side. For the time being, access to the Skolengo section is reserved for the application in its DEV mode.

I've added some todo 'cause the integration is/cannot be finished yet. It's for the following weeks.