refactor(settings/subjects): simplify redundant call to `setOnSubjects`
There was a redundant call to `setOnSubjects` (one if no emoji was provided and another if one or more emoji were provided). This commit simplifies the redundant call by using an external variable. Note that there is a small delay between appending a second emoji and it being selected as the emoji for the subject.